[Refactor #11] Rename package 'entity' to 'domain'
