+++ src/main/java/com/gdg/coffee/cafe/domain/Cafe.java
@@ -1,7 +1,7 @@
-package com.gdg.coffee.cafe.entity;
+package com.gdg.coffee.cafe.domain;
 
 import com.gdg.coffee.domain.common.BaseTime;
-import com.gdg.coffee.member.entity.Member;
+import com.gdg.coffee.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.*;
