+++ src/main/java/com/gdg/coffee/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package com.gdg.coffee.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.gdg.coffee.member.entity.Member;
+import com.gdg.coffee.member.domain.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
